refactor(io): - split convertToSolid like functionality out to convertToSolid module - bulk of convertToBlob is now prepareOutput - added convertToBlob that ONLY deals with the conversion to blobs - small fix in formats for js
